Refactoring password inputs into their own component
